:hammer_and_wrench: Rename files and update genres cards
Update genres anime cards a bit, not a big change :+1:
