+++ src/lib/components/shared/tippies/my_list_anime_details.svelte
